Adds additional capability checks throughout

+++ app/Form/Listeners/ResetUserPreferences.php
@@ -11,6 +11,7 @@ class ResetUserPreferences extends BaseHandler
 
 	private function reset()
 	{
+		if ( !current_user_can('manage_options') ) return;
 		global $wpdb;
 		$wpdb->delete($wpdb->usermeta, ['meta_key' => 'np_visible_posts']);
 		return wp_send_json(['status' => 'success']);
